images/resizeimg.py
from PIL import Image

basewidth = 20
img = Image.open('report.png')
wpercent = (basewidth/float(img.size[0]))
hsize = int((float(img.size[1])*float(wpercent)))
img = img.resize((basewidth,hsize), Image.ANTIALIAS)
img.save('report.png') 