requirements.txt
tkcalendar==1.6.1
fpdf==1.7.2
Pillow